@@ -29,14 +29,14 @@
     ],
     "private": true,
     "dependencies": {
-        "@angular/animations": "^13.3.1",
-        "@angular/common": "^13.3.1",
-        "@angular/compiler": "^13.3.1",
-        "@angular/core": "^13.3.1",
-        "@angular/forms": "^13.3.1",
-        "@angular/platform-browser": "^13.3.1",
-        "@angular/platform-browser-dynamic": "^13.3.1",
-        "@angular/router": "^13.3.1",
+        "@angular/animations": "^14.3.0",
+        "@angular/common": "^14.3.0",
+        "@angular/compiler": "^14.3.0",
+        "@angular/core": "^14.3.0",
+        "@angular/forms": "^14.3.0",
+        "@angular/platform-browser": "^14.3.0",
+        "@angular/platform-browser-dynamic": "^14.3.0",
+        "@angular/router": "^14.3.0",
         "@mdi/js": "^6.6.96",
         "@overwolf/types": "^3.23.1",
         "@popperjs/core": "^2.11.5",
@@ -67,9 +67,9 @@
         "zone.js": "^0.11.5"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^13.3.2",
-        "@angular/cli": "^13.3.3",
-        "@angular/compiler-cli": "^13.3.1",
+        "@angular-devkit/build-angular": "^14.2.11",
+        "@angular/cli": "^14.2.11",
+        "@angular/compiler-cli": "^14.3.0",
         "@types/bootstrap": "^5.1.9",
         "@types/d3": "^7.1.0",
         "@types/express": "^4.17.17",
